requirements.txt
# Core dependencies
google-generativeai>=0.3.0
numpy>=1.24.0
python-dotenv>=1.0.0

# MCP integration
mcp>=0.1.0

# Optional dependencies
pydantic>=2.0.0
tenacity>=8.2.0  # For retry logic

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
